requirements.txt
Flask
requests
openai == 0.28.0
gunicorn
google-cloud-storage
google-auth
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
flask-dance==7.1.0
frozenlist==1.4.1
google-auth==2.17.3
google-cloud-datastore==2.15.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
openai==0.28.0
python-dotenv==1.0.1
rsa==4.9
six==1.16.0
tqdm==4.66.2
urllib3==2.2.1
Werkzeug==3.0.2
yarl==1.9.4
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib